+++ lib/sensu-plugins-kafka.rb
@@ -1,2 +1,0 @@
-require 'sensu-plugins-kafka/common_kafka'
-require 'sensu-plugins-kafka/version'