+++ lib/sensu-plugins-kafka2.rb
@@ -1,0 +1,2 @@
+require 'sensu-plugins-kafka2/common_kafka'
+require 'sensu-plugins-kafka2/version'