+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This CHANGELOG follows the format located [here](https://github.com/sensu-plugin
 
 ## [Unreleased]
 
+### Breaking Changes
+- renamed to `sensu-plugins-kafka2` to avoid namespace conflict per [this](https://github.com/obazoud/sensu-plugins-kafka/issues/21) issue (@majormoses)
+
 ### Added
 - check-kafka-ping.rb(@yuri-zubov sponsored by Actility, https://www.actility.com)
 - check-topic.rb(@yuri-zubov sponsored by Actility, https://www.actility.com)